+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Revision history for muzor
 
+## 0.1.3 -- 2026-7-14
+
+* Error handling, String multiplication
+
+## 0.1.2 -- 2026-7-14
+
+* Expression evaluating
+
 ## 0.1.1 -- 2026-7-14
 
 * Expression parsing